requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
